ga360-bigquery-partitions/requirements.txt
google-cloud-bigquery==1.5.1